Fibonacci type change, Tests

+++ Recursion/main.cpp
@@ -6,7 +6,7 @@ using std::endl;
 void elevator(int floor);
 unsigned long long int Factorial(int n);
 double Power(int n, int power);
-void Fibonacci(unsigned int n, unsigned int a = 0, unsigned int b = 1);
+void Fibonacci(unsigned long long int n, unsigned long long int a = 0, unsigned long long int b = 1);
 
 int main()
 {
@@ -17,18 +17,18 @@ int main()
 	elevator(n);
 
 	cout << "Введите число: "; cin >> n;
-	cout << n << "! = " << Factorial(n) << endl;
+	cout << n << "! = " << Factorial(n) << endl; // максимум: 65! (9223372036854775808)
 
 	int power;
 	cout << "Введите число: "; cin >> n;
 	cout << "Введите степень: "; cin >> power;
 	cout << n << "^" << power << " = " << Power(n, power) << endl;
 
-	unsigned int a;
+	unsigned long long int a;
 	while (true)
 	{
 		cout << "Введите число до которого вывести ряд Фибоначчи: "; cin >> a;
-		Fibonacci(a);
+		Fibonacci(a); // максимальный номер элемента поледовательности: 93 (12200160415121876738)
 		cout << endl;
 	}
 
@@ -58,14 +58,14 @@ double Power(int n, int power)
 	if (power < 0) return (double)1 / Power(n, -power);
 	return n * Power(n, power - 1);
 }
-void Fibonacci(unsigned int n, unsigned int a, unsigned int b)
+void Fibonacci(unsigned long long int n, unsigned long long int a, unsigned long long int b)
 {
 	if (n == 0)
 	{
 		cout << 0;
 		return;
 	}
-	unsigned int c = a + b;
+	unsigned long long int c = a + b;
 	if (c > n) return;
 	if (c == 1) cout << a << ' ' << b << ' ';
 	cout << c << ' ';
